+++ CopilotKit/packages/react-textarea/src/types/autosuggestion-state.ts
@@ -1,6 +1,0 @@
-import { BasePoint } from "slate";
-
-export interface AutosuggestionState {
-  text: string;
-  point: BasePoint;
-}