+++ CopilotKit/packages/react-textarea/src/types/autosuggestions-bare-function.ts
@@ -1,5 +1,0 @@
-export type AutosuggestionsBareFunction = (
-  textBefore: string,
-  textAfter: string,
-  abortSignal: AbortSignal
-) => Promise<string>;