+++ CopilotKit/packages/react-textarea/src/types/index.ts
@@ -1,4 +1,2 @@
-export { defaultBaseAutosuggestionsConfig } from "./base-autosuggestions-config";
-
-export type { BaseAutosuggestionsConfig } from "./base-autosuggestions-config";
-export type { AutosuggestionsBareFunction } from "./autosuggestions-bare-function";
+export * from "./base";
+export * from "./standard-autosuggestions";